// src/app/models/Supplier.ts
export class Supplier {
  kode_supplier: string;
  nama_supplier: string;
  alamat: string;
  kota: string;
  kode_pos: string;
  email: string;
  no_telp_1: string;
  no_telp_2: string;
  no_telp_3: string;
  atas_nama_rekening: string;
  no_rekening: string;
  catatan: string;
  kode_top: string;
}